python-service/requirements.txt
requests==2.26.0
Pillow==9.0.0